chore: Integrate home pages into Planner
