+++ pages/_app.tsx
@@ -6,12 +6,10 @@ import { createTheme, StyledEngineProvider, ThemeProvider } from '@mui/material/
 import { AnimateSharedLayout } from 'framer-motion';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
-import { useRouter } from 'next/router';
 import React, { useEffect } from 'react';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
 
-import TopAndSidebar from '../components/common/Top-and-SideBar/TopAndSidebar';
 import { AuthProvider } from '../modules/auth/auth-context';
 import { useStore } from '../modules/redux/store';
 
@@ -60,41 +58,6 @@ if (!firebase.apps.length) {
 }
 
 /**
- * Determines if the page at a specific route needs the AppNavigation component
- * @param pathname page route
- * @returns boolean
- */
-const needAppNav = (pathname: string): boolean => {
-  const routesList = ['/app/onboarding', '/auth', '/app/plans/', '/app/test'];
-  if (pathname === '/') {
-    return false;
-  }
-  for (let i = 0; i < routesList.length; i++) {
-    if (pathname.startsWith(routesList[i])) {
-      return false;
-    }
-  }
-  return true;
-};
-
-/**
- * Renders page layout for all pages
- * TODO: Consider unifying all NavigationBars here
- */
-function PageLayout({ Component, pageProps }) {
-  const router = useRouter();
-
-  const content = needAppNav(router.pathname) ? (
-    <TopAndSidebar>
-      <Component {...pageProps} />
-    </TopAndSidebar>
-  ) : (
-    <Component {...pageProps} />
-  );
-  return content;
-}
-
-/**
  * The root wrapper component for the app.
  *
  * This initializes the app's authentication and the app root store.
@@ -134,9 +97,10 @@ export default function NebulaApp({ Component, pageProps }: AppProps): JSX.Eleme
       <Provider store={store}>
         <PersistGate loading={<p>Loading...</p>} persistor={persistor}>
           <AnimateSharedLayout>
-            {' '}
             <StyledEngineProvider injectFirst>
-              <ThemeProvider theme={theme}>{PageLayout({ Component, pageProps })}</ThemeProvider>
+              <ThemeProvider theme={theme}>
+                <Component {...pageProps} />
+              </ThemeProvider>
             </StyledEngineProvider>
           </AnimateSharedLayout>
         </PersistGate>
